requirements.txt
python-vlc
libusb
pyqt5
PyQtWebEngine
requests
onvif_zeep
opencv-python